requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrdict==2.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
discord==1.0.1
discord.py==1.3.4
dispander==0.4.0
distlib==0.3.1
echidna==0.2.1
filelock==3.0.12
idna==2.9
multidict==4.7.5
numpy==1.19.0
pipenv==2020.6.2
pycparser==2.20
PyNaCl==1.4.0
redis==3.5.1
six==1.14.0
Pillow==8.1.0
virtualenv==20.0.26
virtualenv-clone==0.5.4
websockets==8.1
yarl==1.4.2